refactor: remove old REST types
