+++ apps/admin-ui/src/component/StatusCell.tsx
@@ -6,56 +6,7 @@ import {
   ApplicationEventStatusChoice,
   ApplicationStatusChoice,
 } from "common/types/gql-types";
-
-const getApplicationStatusColor = (
-  status: ApplicationStatusChoice,
-  size: "s" | "l"
-): string => {
-  switch (status) {
-    case ApplicationStatusChoice.Received:
-    case ApplicationStatusChoice.Draft:
-      return "var(--color-info)";
-    case ApplicationStatusChoice.Handled:
-    case ApplicationStatusChoice.InAllocation:
-      return "var(--color-success)";
-    case ApplicationStatusChoice.ResultsSent:
-      return "var(--color-white)";
-    case ApplicationStatusChoice.Expired:
-    case ApplicationStatusChoice.Cancelled:
-    default:
-      switch (size) {
-        case "s":
-          return "var(--color-error)";
-        case "l":
-        default:
-          return "var(--color-error-dark)";
-      }
-  }
-};
-
-const getApplicationEventStatusColor = (
-  status: ApplicationEventStatusChoice,
-  size: "s" | "l"
-): string => {
-  switch (status) {
-    case ApplicationEventStatusChoice.Reserved:
-    case ApplicationEventStatusChoice.Unallocated:
-      return "var(--color-info)";
-    case ApplicationEventStatusChoice.Approved:
-      return "var(--color-success)";
-    // return "var(--color-alert-light)";
-    case ApplicationEventStatusChoice.Failed:
-    case ApplicationEventStatusChoice.Declined:
-    default:
-      switch (size) {
-        case "s":
-          return "var(--color-error)";
-        case "l":
-        default:
-          return "var(--color-error-dark)";
-      }
-  }
-};
+import { getApplicationEventStatusColor, getApplicationStatusColor } from "./applications/util";
 
 const StatusDot = styled.div<{
   status: ApplicationStatusChoice;
